refactor(server): rename userAction module file
